NS-Toolbox-4 Cloning currently does not work on non primitive fields.
- Tests not more clearly doubles as usage documentation.
- Added builder case.
